Extract `Bookkeeper` class definition into own module
